requirements.txt
opencv-python>=4.11.0.86
tqdm>=4.67.1
pillow>=11.1.0
numpy>=2.3.2
supervision>=0.26.1
